refactor(workspace)!: absorb the control-plane shell into the API crate
scylla-control-plane was a 200-line crate holding nothing but a main.rs, a
config wrapper and a 60-line runtime. Every one of ControlPlaneConfig's four
methods delegated straight to scylla_api::CoreConfig, and its ConfigError was
a character-for-character copy of scylla_api::error::ConfigError. The crate
boundary bought nothing and cost a manifest, a feature-forwarding hop
(register = ["scylla-api/register"]) and a duplicated error type.

The shell now lives in the crate it was wrapping, which is renamed
scylla-control-plane after the artifact it ships. CoreConfig becomes
ControlPlaneConfig: the wrapper flattened it with #[serde(flatten)], so
removing the indirection leaves the TOML shape byte-identical and the three
shipped config files untouched.

The merge order matters and is deliberate: move the binary and the config
directory into the API crate first, delete the shell, then rename the
directory last. That keeps `package.name == [[bin]].name ==
"scylla-control-plane"` and the `crates/scylla-control-plane/config/` path
true at the commit boundary, so Dockerfile, docker-compose.yaml and justfile
need no edit at all.

Workspace goes from 5 crates to 4.

Verified beyond `cargo check`: the binary builds under its expected name,
boots against a live Postgres, applies migrations, bootstraps the admin,
loads the Cedar policy set, serves gRPC on 50051 and shuts down cleanly on
SIGTERM. `docker compose config` still resolves.
