--- crates/scylla-api/Cargo.toml
@@ -1,48 +1,0 @@
-[package]
-name = "scylla-api"
-version = "0.1.0"
-edition = "2024"
-
-[features]
-# Public self-service signup (the gRPC RegistrationService). Off by default so a
-# deployment stays invite-only / OAuth-only unless it opts in.
-register = []
-
-[dependencies]
-# shared
-scylla-core = { workspace = true, features = ["postgres", "hash", "permission"] }
-tokio.workspace = true
-async-trait.workspace = true
-tower.workspace = true
-tower-http.workspace = true
-http.workspace = true
-serde.workspace = true
-toml.workspace = true
-chrono.workspace = true
-prost-types.workspace = true
-tracing.workspace = true
-derive_more.workspace = true
-thiserror.workspace = true
-
-tokio-stream = { workspace = true }
-futures-util.workspace = true
-serde_json.workspace = true
-
-# webhook ingress HTTP server (separate listener from the gRPC port)
-axum.workspace = true
-
-# database
-sqlx.workspace = true
-
-# grpc
-tonic.workspace = true
-tonic-web.workspace = true
-tonic-reflection.workspace = true
-tonic-async-interceptor.workspace = true
-scylla-protocol.workspace = true
-
-[dev-dependencies]
-tokio = { workspace = true }
-
-[lints]
-workspace = true
