+++ crates/scylla-control-plane/src/bootstrap.rs
