+++ crates/scylla-control-plane/src/error.rs
