+++ crates/scylla-control-plane/src/grpc/handlers/agent_handler.rs
