+++ crates/scylla-control-plane/src/grpc/handlers/app_auth_handler.rs
