+++ crates/scylla-control-plane/src/grpc/handlers/auth_handler.rs
