+++ crates/scylla-control-plane/src/grpc/handlers/grant_handler.rs
