+++ crates/scylla-control-plane/src/grpc/handlers/invitation_handler.rs
