+++ crates/scylla-control-plane/src/grpc/handlers/macros.rs
