+++ crates/scylla-control-plane/src/grpc/handlers/pipeline_handler.rs
