+++ crates/scylla-control-plane/src/grpc/handlers/project_handler.rs
