+++ crates/scylla-control-plane/src/grpc/handlers/registration_handler.rs
