+++ crates/scylla-control-plane/src/grpc/handlers/secret_handler.rs
