+++ crates/scylla-control-plane/src/grpc/mappers/error_mapper.rs
