+++ crates/scylla-control-plane/src/grpc/mappers/job_log_mapper.rs
