+++ crates/scylla-control-plane/src/grpc/mappers/organization_mapper.rs
