+++ crates/scylla-control-plane/src/grpc/mappers/pipeline_mapper.rs
