+++ crates/scylla-control-plane/src/grpc/mappers/secret_mapper.rs
