+++ crates/scylla-control-plane/src/grpc/mappers/user_mapper.rs
