+++ crates/scylla-control-plane/src/grpc/middleware/auth_interceptor.rs
